feat: Allow disabling exit on error in finalize()
This commit introduces a new optional parameter `notExit` to the `finalize()` function. When `notExit` is set to true, the program will not exit after printing error messages, allowing the user to handle the errors and continue execution. The function now returns true if there are no error messages, and false if there are any and `notExit` is true.